% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ApproximateEntropy}
\alias{ApproximateEntropy}
\title{Calculate the approximate entropy of a time series with this function.}
\usage{
ApproximateEntropy(x, dim, R)
}
\arguments{
\item{x}{is the time series to analyse}

\item{dim}{is the embedding dimension of the time series}

\item{R}{is the radius in which to search for matches}
}
\description{
Calculate the approximate entropy of a time series with this function.
}
