% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppsr}
\alias{cppsr}
\title{Constant Embedding Parameters and Principal Component Analysis-based Phase Space Reconstruction}
\usage{
cppsr(x, m)
}
\arguments{
\item{x}{A single column time series}

\item{m}{An integer indicating the embedding dimension of the time series}
}
\value{
The output of the algorithm is a list that includes:
\itemize{
 \item \code{Yprime} The reconstructed attractor of the time series.
}
}
\description{
Reconstruct attractor based on constant embedding parameters and principle component analysis.
}
\details{
Constant embedding parameters and Principal component analysis-based Phase Space Reconstruction (CPPSR) is a specific version of phase space reconstruction that is applicable to low-dimensional systems (dimension < 3).
}
\examples{
# Generate example time series data and inputs for FNN
t = seq(0, 1, 0.01)
x = sin(2*pi*10*t) + 2*cos(2*pi*5*t)
maxDim = 10
delay = 1 # tau for CPPSR is fixed to tau = 1
rtol = 10
atol = 15
fnn_tol = 0.01

# Run FNN
fnn_out = false_nearest_neighbors(x, maxDim = maxDim, delay = delay, rtol = rtol, 
                                  atol = atol, fnn_tol = fnn_tol)

# Run CPPSR
y_cppsr = cppsr(x = x, m = fnn_out$dim)

}
\references{
Li, D., Cao, M., Manoach, E. & Ragulskis, M. A novel embedding method for characterization of low-dimensional nonlinear dynamical systems. Nonlinear Dyn 104, 125–148 (2021). https://doi.org/10.1007/s11071-021-06229-1
}
