% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Ent_Ap}
\alias{Ent_Ap}
\title{Approximate Entropy}
\usage{
Ent_Ap(x, dim, R)
}
\arguments{
\item{x}{is the time series to analyse}

\item{dim}{is the embedding dimension of the time series}

\item{R}{is the radius in which to search for matches}
}
\value{
The output of the algorithm is a single value that reflects the entropy of the time series in bits.
}
\description{
Calculate the approximate entropy of a time series.
}
\details{
Approximate entropy is a measure of a systems regularity or predictability. First introduced in 1991 by Steven Pincus as a way to quantify the predictability of heart rate variability, it has since been used in a wider physiology research and even extended to economics. Approximate entropy is aims to overcome the effects of noise in the time series.

The steps in the algorithm are generally rather simple. The time series is divided up into vectors of length \eqn{m}. Each vector is then compared to the original time series as well as all other vectors with matches to similar ones counted. In this function Euclidean distance is used to match vector within the tolerance specified by \eqn{r}. The log average of the counts is taken and the process is repeated for vectors of length \eqn{m+1}. The final step is subtracting the conditional probability of \eqn{m + 1} from that of \eqn{m}. 

Approximate entropy is measured in bits and is bound between 0 bits (entirely predictable) to 2 bits (completely random). Approximate entropy has been found to be inconsistent in its results, very data hungry and also contains a bias towards more predictability as a result of the self-match.
}
\examples{

x = rnorm(1000)
dim = 8
R = 0.2

AE = Ent_Ap(x, dim, R)
}
\references{
Pincus, S. M. (1991). Approximate entropy as a measure of system complexity. Proceedings of the National Academy of Sciences, 88(6), 2297–2301. https://doi.org/10.1073/pnas.88.6.2297

Richman, J. S., & Moorman, J. R. (2000). Physiological time-series analysis using approximate entropy and sample entropy. American journal of physiology-heart and circulatory physiology, 278(6), H2039-H2049.

Yentes, J. M., & Raffalt, P. C. (2021). Entropy Analysis in Gait Research: Methodological Considerations and Recommendations. Annals of Biomedical Engineering, 49(3), 979–990. https://doi.org/10.1007/s10439-020-02616-8
}
