% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Ent_Sym}
\alias{Ent_Sym}
\title{Symbolic Entropy}
\usage{
Ent_Sym(x, thresholdVal, seqLength)
}
\arguments{
\item{x}{A single column time series}

\item{thresholdVal}{The threshold to decide the symbol (0 or 1) of each value. Typically, 0, mean, or median of the time series is chosen.}

\item{seqLength}{The length of the vectors to be compared for matches}
}
\value{
The output of the algorithm is a single value that reflects the entropy of the time series in bits.
}
\description{
Calculate the symbolic entropy of a time series.
}
\details{
Like all entropy functions, this one also quantifies the amount of complexity (or uncertainty/unpredictability) in the signal. Higher values indicates that the time series is less predictable (or more random) and the temporal patterns in the time series are less repetitive. Lower values indicates that the time series is more predictable and the temporal patterns in the time series are more repetitive. Symbolic entropy is rather robust to the length of the time series, making it more suitable for short time series compared to other entropy functions.
}
\examples{

x = rnorm(1000)
thresholdVal = mean(x)
seqLength = 2

SymE = Ent_Sym(x, thresholdVal, seqLength)

}
\references{
Aziz, W., Arif, M. Complexity analysis of stride interval time series by threshold dependent symbolic entropy. Eur J Appl Physiol 98, 30–40 (2006). https://doi.org/10.1007/s00421-006-0226-5
}
