% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false_nearest_neighbors.R
\name{false_nearest_neighbors}
\alias{false_nearest_neighbors}
\title{False Nearest Neighbors}
\usage{
false_nearest_neighbors(data, maxDim, delay, rtol, atol, fnn_tol)
}
\arguments{
\item{data}{Time series data (numeric vector)}

\item{maxDim}{Maximum embedding dimension to consider}

\item{delay}{Time delay for the embedding}

\item{rtol}{Relative tolerance for the neighbor distance (often, 10)}

\item{atol}{Absolute tolerance for the neighbor distance (often, between 10 and 20, typically 15)}

\item{fnn_tol}{Proportion of false neighbors below which to choose dim}
}
\value{
A list containing the following elements:
\describe{
  \item{fnn}{Percentage of false nearest neighbors for each dimension}
  \item{dim}{Embedding dimension at which the percentage of FNN drops below fnn_tol}
}
}
\description{
Computes the false nearest neighbors for a given time series.
}
\details{
FNN is part of the phase space reconstruction step that is needed for some nonlinear analysis methods.
}
\examples{
# Generate example time series data
x = rnorm(1000)

# Function paramters
maxDim = 10
delay = 1 # You can get this value like: ami_out$tau[1,1]
rtol = 10
atol = 15
fnn_tol = 0.01

# Compute false nearest neighbors
fnn_out = false_nearest_neighbors(x, maxDim = maxDim, delay = delay, rtol = rtol, atol = atol, fnn_tol = fnn_tol)


}
