% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fgn_sim}
\alias{fgn_sim}
\title{Fractional Gaussian Noise Simulation}
\usage{
fgn_sim(n, H, mean = 0, std = 1)
}
\arguments{
\item{n}{The length of the resulting time series}

\item{H}{The Hurst value of the resulting time series. This value should be between 0.01 and 0.99}

\item{mean}{The desired mean of the resulting time series. Default is 0.}

\item{std}{The desired standard deviation of the resulting time series. Default is 1.}
}
\value{
The output of the algorithm is a numeric vector of length \code{n}.
}
\description{
Function to generate Fractional Gaussian Noise
}
\details{
Create a time series with a specified Hurst value.
}
\examples{

ts_out = fgn_sim(n = 1000, H = 0.7)
}
