% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fnn}
\alias{fnn}
\title{False Nearest Neighbor}
\usage{
fnn(x, tau, mmax, rtol, atol)
}
\arguments{
\item{x}{- a single column time series}

\item{tau}{- the first local minimum from the data. This is a value returned from the \code{ami} function}

\item{mmax}{- The maximum embedding dimension. Common practice is to set this to 12.}

\item{rtol}{- The near tolerance. Common practice is to set this to 15.}

\item{atol}{- The far tolerance. Used to accommodate for the fact that all neighbors may be far away to begin with. Common practice is to set this to 2.}
}
\value{
The output of the algorithm is a list that includes:
\itemize{
 \item \code{dE} A column vector of the percentages of false neighbors at the embedding dimensions up to \code{mmax}
 \item \code{dim} An integer reflecting the time delay at which there is the lowest number of false neighbors
}
}
\description{
Calculate the average mutual information of a time series.
}
\details{
AMI is part of the phase space reconstruction step that is needed for some nonlinear analysis methods.
}
\examples{

x = rnorm(1000)
tau = 3 # You can get this value like: ami_out$tau[1,1]
mmax = 12
rtol = 15
atol = 2

fnn_out = fnn(x = x, tau = tau, mmax = mmax, rtol = rtol, atol = atol)


}
