% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{healthy_young}
\alias{healthy_young}
\title{Gaitprint data for a healthy young person}
\format{
A data frame with 10,000 rows and 18 columns:
\describe{
  \item{time}{Time in seconds. Data is sampled at 200Hz}
  \item{thigh_pitch_lt_deg}{Left thigh angle in reference to the direction of heading}
  \item{shank_pitch_lt_deg}{Left shank angle in reference to the direction of heading}
  \item{contact_lt}{Pre calculated heel contact events for the left leg. 0 means the foot is in the air and 1000 indicates contact with the ground}
  \item{foot_pitch_lt_deg}{Left foot angle in reference to the direction of heading}
  \item{thigh_pitch_rt_deg}{Right thigh angle in reference to the direction of heading}
  \item{shank_pitch_rt_deg}{Right shank angle in reference to the direction of heading}
  \item{contact_rt}{Pre calculated heel contact events for the right leg. 0 means the foot is in the air and 1000 indicates contact with the ground}
  \item{foot_pitch_rt_deg}{Right foot angle in reference to the direction of heading}
  \item{lumbar_flexion_deg}{Lumbar spine angle in reference to the direction of heading}
  \item{hip_flexion_lt_deg}{Left hip angle in reference to the direction of heading}
  \item{knee_flexion_lt_deg}{Left knee angle in reference to the direction of heading}
  \item{ankle_dorsiflexion_lt_deg}{Left ankle angle in reference to the direction of heading}
  \item{hip_flexion_rt_deg}{Right hip angle in reference to the direction of heading}
  \item{knee_flexion_rt_deg}{Right knee angle in reference to the direction of heading}
  \item{ankle_dorsiflexion_rt_deg}{Right ankle angle in reference to the direction of heading}
}
}
\usage{
healthy_young
}
\description{
A subset of data from the NONAN Gaitprint data repository. The age range for healthy young people is 19-34 years old. For details on the collections and data descriptions see the paper cited in the references. To get the full versions of the data please visit \url{https://springernature.figshare.com/collections/NONAN_GaitPrint_An_IMU_gait_database_of_healthy_young_adults/6415061/1}.
}
\references{
\url{https://www.nature.com/articles/s41597-023-02704-z}
}
\keyword{datasets}
