% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lye_r}
\alias{lye_r}
\title{Lyapunov Rosenstein Method}
\usage{
lye_r(x, tau, dim, fs)
}
\arguments{
\item{x}{- A single column time series.}

\item{tau}{- The first local minimum from the data. This is a value returned from the \code{ami} function.}

\item{dim}{- An integer reflecting the embedding dimension at which there is the lowest number of false neighbors. This is a value returned from the \code{fnn} function}

\item{fs}{- The sampling frequency that the data was collected at.}
}
\value{
The output of the algorithm is a list that includes:
\itemize{
 \item \code{M} The number of matched pairs
 \item \code{out} A matrix of all the matched pairs and the average line divergence from which the slope is calculated. The matched pairs are columns 1 and 2. The average line divergence is column 3.
}
}
\description{
Calculate the average mutual information of a time series.
}
\details{
AMI is part of the phase space reconstruction step that is needed for some nonlinear analysis methods.
}
\examples{

x = rnorm(1000)
tau = 3 # You can get this value like: ami_out$tau[1,1]
dim = 4 # You can get this value like: fnn_out$dim
fs = 60

lye_out = lye_r(x = x, tau = tau, dim = dim, fs = fs)

}
