% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lye_rosenstein}
\alias{lye_rosenstein}
\title{Lyapunov Rosenstein Method}
\usage{
lye_rosenstein(X, samp_rate, mean_freq, nsteps, regpoints)
}
\arguments{
\item{mean_freq}{- A double indicating the mean frequency of the time series.}

\item{nsteps}{- Number of time steps to consider for calculating the Lyapunov exponent.}

\item{regpoints}{- Number of points to include in fitting a line on the divergence plot.}

\item{x}{- A single column time series.}

\item{samp_rate-}{A double indicating the sampling rate of the time series.}
}
\value{
The output of the algorithm is a list that includes:
\itemize{
 \item \code{lye} The Lyapunov exponent of the time series.
 \item \code{time_steps} A vector of the indices of the time steps used to calculate the Lyapunov exponent.
 \item \code{mean_distances} A vector of mean log distances used to calculate the Lyapunov exponent.
}
}
\description{
Calculate Lyapunov exponent of a time series.
}
\details{
Lyapunov exponent is the rate of divergence between two neighboring points of a trajectory in a state space.
}
\examples{

x = rnorm(1000)
tau = 3 # You can get this value like: ami_out$tau[1,1]
dim = 4 # You can get this value like: fnn_out$dim
fs = 60

lye_out = lye_rosenstein(x = x, tau = tau, dim = dim, fs = fs)

}
