% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NONANr-package.R
\docType{package}
\name{nonanR-package}
\alias{nonanR}
\alias{nonanR-package}
\title{nonanR: Nonlinear Analysis Methods for Time Series Analysis}
\description{
This is a collection of functions that can be used for the nonlinear analysis of time series data.
}
\keyword{internal}
