% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ami.R
\name{plot_ami}
\alias{plot_ami}
\title{Average Mutual Information Plot}
\usage{
plot_ami(x)
}
\arguments{
\item{x}{The list object returned \code{ami} in this package.}
}
\description{
Plot method for average mutual information.
}
\examples{

# Create a time series and perform DFA
x = fgn_sim(n = 1000, H = 0.9)
L = 1
bins = 30

ami_out = ami(x, x, L, bins)

# Plot ami_out
plot_ami(ami_out)

}
