% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bayesh.R
\name{plot_bayesh}
\alias{plot_bayesh}
\title{Hurst Exponent Posterior Distribution Plot}
\usage{
plot_bayesh(x)
}
\arguments{
\item{x}{The vector returned \code{bayesH} in this package.}
}
\description{
Plot the posterior distribution of the Hurst exponents sampled from bayesH
}
\examples{
x = fgn_sim(n = 128, H = 0.9)

h.pdf = bayesH(x = x, n = 200) 

# Plot h.pdf
plot_bayesh(h.pdf)

}
