% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dfa.R
\name{plot_dfa}
\alias{plot_dfa}
\title{Detrended Fluctuation Analysis Plot}
\usage{
plot_dfa(x)
}
\arguments{
\item{x}{The list object returned \code{dfa} in this package.}
}
\description{
Plot method for monofractal detrended fluctuation analysis
}
\examples{

# Create a time series and perform DFA
x = rnorm(1000)
order = 1
verbose = 1
scales = c(16,32,64,128,256,512)
scale_ratio = 2

dfa_out = dfa(x, order, verbose, scales, scale_ratio)

# Plot dfa_out
plot_dfa(dfa_out)

}
