% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fnn.R
\name{plot_fnn}
\alias{plot_fnn}
\title{False Nearest Neighbors Plot}
\usage{
plot_fnn(x)
}
\arguments{
\item{x}{The list object returned \code{fnn} in this package.}
}
\description{
Plot method for false nearest neighbors.
}
\examples{

# When performing phase space reconstruction there are two crucial steps. 
# - The first step is running the ami function. 

#' # Create a time series
x = fgn_sim(n = 1000, H = 0.9)

# Specify the parameters for ami
y = x
L = 50
bins = 30

ami_out = ami(x, y, L, bins)

# - The second step is running the fnn function
# Specify the parameters for fnn
tau = 3 # You can get this value like: ami_out$tau[1,1]
mmax = 12
rtol = 15
atol = 2

fnn_out = fnn(x = x, tau = tau, mmax = mmax, rtol = rtol, atol = atol)

# Plot fnn_out
plot_fnn(fnn_out)

}
