% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_iaafft.R
\name{plot_iaafft}
\alias{plot_iaafft}
\title{IAAFFT Surrogates Plot}
\usage{
plot_iaafft(x, surrogates)
}
\arguments{
\item{x}{Original time series}

\item{surrogates}{IAAFFT surrogates}
}
\description{
Plot method for IAAFFT surrogates.
}
\examples{
# Generate example time series data
x = rnorm(1000)

# IAAFFT surrogates
iaafft_out = iaafft(x, N = 19)

# Plot
plot_iaafft(x, iaafft_out)

}
