% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_iaaft.R
\name{plot_iaaft}
\alias{plot_iaaft}
\title{IAAFT Surrogates Plot}
\usage{
plot_iaaft(x, surrogates)
}
\arguments{
\item{x}{Original time series}

\item{surrogates}{IAAFT surrogates}
}
\description{
Plot method for IAAFT surrogates.
}
\examples{

# Original time series
x = rnorm(1000)

# IAAFT surrogates
iaaft_out = nonanR::iaafft(x, N = 19)
# Plot
plot_iaaft(x, iaaft_out)

}
