% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lye.R
\name{plot_lye}
\alias{plot_lye}
\title{Log Divergence Plot}
\usage{
plot_lye(x)
}
\arguments{
\item{x}{The list object returned \code{lye_rosenstein} in this package.}
}
\description{
Plot method for log of divergence.
}
\examples{
fs = 100
t = seq(0, 10, 1/fs)
ts = sin(2*pi*10*t) + 2*cos(2*pi*5*t)

mean_frequency = meanfreq(signal = ts, samp_rate = fs)

mmax = 12
rtol = 15
atol = 2

time_delay = ami(ts, ts, 50, 30)
tau = time_delay$tau[1,1] # Optimal time delay estimated by AMI

embed = fnn(x = ts, tau = tau, mmax = mmax, rtol = rtol, atol = atol)
dim = embed$dim # Optimal embedding dimension estimated by FNN

psr_length = length(ts) - tau*(dim-1)
start = 1
stop = psr_length
X = matrix(nrow = psr_length, ncol = dim)
for (i in 1:dim) {
  X[,i] = ts[start:stop]
  start = start + tau
  stop = stop + tau
}

lye_out = lye_rosenstein(X = X, samp_rate = fs, mean_freq = mean_frequency, 
                         nsteps = 500, regpoints = 10:500)


# Plot lye_out
plot_lye(lye_out)

}
