% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lye_r.R
\name{plot_lye_r}
\alias{plot_lye_r}
\title{Lyapunov Exponent - Rosenstein Plot}
\usage{
plot_lye_r(x)
}
\arguments{
\item{x}{The list object returned \code{LyE_R} in this package.}
}
\description{
Plot method for Lyapunov Exponent - Rosenstein.
}
\examples{

# Create a time series
x = rnorm(1000)
tau = 3 # You can get this value like: ami_out$tau[1,1]
dim = 4 # You can get this value like: fnn_out$dim
fs = 60

lye_out = lye_r(x = x, tau = tau, dim = dim, fs = fs)

# Plot ami_out
plot_lye_r(lye_out)

}
