% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rqa.R
\name{plot_rqa}
\alias{plot_rqa}
\title{Plot RQA Recurrence Matrix}
\usage{
plot_rqa(recurrence_matrix)
}
\arguments{
\item{recurrence_matrix}{The output from the \code{rqa} function}
}
\description{
This function allows you to use the NONAN theme for ggplot2 graphics
}
\examples{

library(nonanR)

# Create a sample time series
x = fgn_sim(n = 100, H = 0.8)

# Compute RQA
x.recpt = rqa(x, x, 1, 1, 0, 1, 2, 2, 0, .0001, 0, 1)

# Return recurrence plot
plot_rqa(x.recpt)


}
