% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Surr_PseudoPeriodic.R
\name{plot_Surr_PseudoPeriodic}
\alias{plot_Surr_PseudoPeriodic}
\title{IAAFFT Surrogates Plot}
\usage{
plot_Surr_PseudoPeriodic(x, surrogate)
}
\arguments{
\item{x}{Original time series}

\item{surrogate}{Pseudoperiodic surrogate}
}
\description{
Plot method for Pseudoperiodic surrogates.
}
\examples{
# Generate example time series data
x <- sin(seq(0, 2 * pi, length.out = 500))

# Pseudoperiodic surrogate
PseudoPeriodic_out  <- Surr_PseudoPeriodic(x, tau, dim, rho)

# Plot
plot_Surr_PseudoPeriodic(x, PseudoPeriodic_out$ys)

}
