% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SampleEntropy}
\alias{SampleEntropy}
\title{Sample Entropy}
\usage{
SampleEntropy(x, m, R)
}
\arguments{
\item{x}{A single column time series}

\item{m}{The length of the vectors to be compared for matches}

\item{R}{The radius for accepting matches}
}
\value{
The output of the algorithm is a single integer that reflects the entropy of the time series in bits.
}
\description{
Calculate the sample entropy of a time series.
}
\details{
Sample entropy can be used to quantify the complexity of a time series. A higher sample entropy value is associated with a more complex signal (more random). On the other hand a time series with a lower sample entropy value could be considered more predictable and repetitive.
}
\examples{

x = rnorm(1000)
m = 2
R = 0.2

SE = SampleEntropy(x, m, R)

}
\references{
Richman, J.S., Moorman, J.R., 2000. Physiological time-series analysis using approximate entropy and sample entropy. Am. J. Physiol. Heart Circ. Physiol. 278. https://doi.org/10.1152/ajpheart.2000.278.6.H2039
}
