% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nonan.R
\name{theme_nonan}
\alias{theme_nonan}
\title{This function allows you to use the NONAN theme for ggplot2 graphics}
\usage{
theme_nonan(background = "#646666", text = "#D9D9D9", axes = "#8A8A8A")
}
\description{
This function allows you to use the NONAN theme for ggplot2 graphics
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(x = mpg, y = disp, color = cyl)) + 
  geom_point(aes(colour = factor(cyl))) + 
  theme_nonan()
  
  
# Alternatively you could provide color values for the different features of the graph
ggplot(mtcars, aes(x = mpg, y = disp, color = cyl)) + 
  geom_point(aes(colour = factor(cyl))) + 
  theme_nonan(background = "#000000", text = "#ffffff", axes = "#888888")
  

}
