% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ami}
\alias{ami}
\title{Average Mutual Information}
\usage{
ami(x, y, L, bins)
}
\arguments{
\item{x}{- a single column time series}

\item{y}{- a single column time series. This can be the same as x}

\item{L}{- the maximum lag of the time series. This is usually the same as the sampling frequency.}

\item{bins}{- the number of histogram bins to split the data into. You can specify 0 and the algorithm will bin the data for you.}
}
\value{
The output of the algorithm is a list that includes:
\itemize{
 \item \code{tau} A data frame of the local minima of the AMI values and the corresponding lag
 \item \code{ami} A data frame of all the AMI values at each lag
}
}
\description{
Calculate the average mutual information of a time series.
}
\details{
AMI is part of the phase space reconstruction step that is needed for some nonlinear analysis methods.
}
\examples{

x = rnorm(1000)
y = x
L = 50
bins = 30 # If you do not want to specify a bin number, you can set it to 0.

ami_out = ami(x, y, L, bins)


}
