% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AMI}
\alias{AMI}
\title{Average Mutual Information}
\usage{
AMI(x, L)
}
\arguments{
\item{x}{- a single column time series}

\item{L}{- the maximum lag of the time series. This is usually the same as the sampling frequency.}
}
\description{
Calculate the average mutual information of a time series.
}
\details{
AMI is part of the phase space reconstruction step that is needed for some nonlinear analysis methods.
}
\examples{

x = rnorm(1000)
L = 100

ami_out = AMI(x, L)


}
