% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppsr}
\alias{cppsr}
\title{Constant Embedding Parameters and Principal Component Analysis-based Phase Space Reconstruction}
\usage{
cppsr(x, m)
}
\arguments{
\item{x}{- A single column time series.}

\item{m}{- An integer indicating the embedding dimension of the time series.}
}
\value{
The output of the algorithm is a list that includes:
\itemize{
 \item \code{Yprime} The reconstructed attractor of the time series.
}
}
\description{
Reconstruct attractor based on constant embedding parameters and principle component analysis.
}
\details{
Constant embedding parameters and principal component analysis-based phase space reconstruction (CPPSR) is a specific version of phase space reconstruction that is applicable on low-dimensional systems (dimension < 3).
}
\examples{

x = sin(2*pi*10) + 2*cos(2*pi*5)

Yprime = cppsr(x = x, m = 2)

}
