% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Ent_Ap}
\alias{Ent_Ap}
\title{Approximate Entropy}
\usage{
Ent_Ap(x, dim, R)
}
\arguments{
\item{x}{is the time series to analyse}

\item{dim}{is the embedding dimension of the time series}

\item{R}{is the radius in which to search for matches}
}
\value{
The output of the algorithm is a single integer that reflects the entropy of the time series in bits.
}
\description{
Calculate the approximate entropy of a time series.
}
\details{
Here are some additional details about approximate entropy.
}
\examples{

x = rnorm(1000)
dim = 8
R = 0.2

AE = Ent_Ap(x, dim, R)
}
