% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Ent_Sym}
\alias{Ent_Sym}
\title{Symbolic Entropy}
\usage{
Ent_Sym(x, thresholdVal, seqLength)
}
\arguments{
\item{x}{A vector of integers}

\item{thresholdVal}{The threshold of the search that you want to do}

\item{seqLength}{The length of the sequence that you want to find}
}
\value{
The output of the algorithm is a single integer that reflects the entropy of the time series in bits.
}
\description{
Calculate the symbolic entropy of a time series.
}
\details{
Like all entropy functions, this one also quantifies the amount of complexity (or uncertainty/unpredictability) in the signal. Symbolic entropy can be used to gather greater insight into the underlying patterns seen in movement data.
}
\examples{

x = rnorm(1000)
thresholdVal = 2
seqLength = 0.2

SymE = Ent_Sym(x, thresholdVal, seqLength)

}
\references{
Aziz, W., Arif, M. Complexity analysis of stride interval time series by threshold dependent symbolic entropy. Eur J Appl Physiol 98, 30–40 (2006). https://doi.org/10.1007/s00421-006-0226-5
}
