% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false_nearest_neighbors.R
\name{false_nearest_neighbors}
\alias{false_nearest_neighbors}
\title{False Nearest Neighbors}
\usage{
false_nearest_neighbors(data, maxDim, delay, rtol, atol, fnn_tol)
}
\arguments{
\item{data}{Time series data (numeric vector)}

\item{maxDim}{Maximum embedding dimension to consider}

\item{delay}{Time delay for the embedding}

\item{rtol}{Relative tolerance for the neighbor distance (often, 10)}

\item{atol}{Absolute tolerance for the neighbor distance (often, between 10 and 20, typically 15)}

\item{fnn_tol}{Proportion of false neighbors below which to choose dim}
}
\value{
A list containing the following elements:
\describe{
  \item{fnn}{Percentage of false nearest neighbors for each dimension}
  \item{dim}{Embedding dimension at which the percentage of FNN drops below fnn_tol}
}
}
\description{
Computes the false nearest neighbors for a given time series.
}
\details{
Finding the embedding dimensions is typically the second step in the phase space reconstruction. This can be done using the False Nearest Neighbors which minimizes the number of false neighbors as the attractor (input time series) is unfolded at increasingly higher and higher dimensions. In the package, FNN follows the calculation of average mutual information (AMI), where the output tau serves as an input into this function. 

The False Nearest Neighbors algorithm calculates the distance between a vector and its nearest neighbor in an m-dimensional space, repeating the process for each subsequent dimension by incrementing m by 1. True neighbors exhibit minimal change in distance between dimensions d and d+1. The algorithm then computes the ratio of the difference in distances between vectors at dimensions d and d+1 to the distance at dimension d. If this ratio exceeds a predefined threshold, `rtol`, the vector is determined to be a false neighbor. As the dimension increases, the percentage of false neighbors should decrease as the dynamics of the attractor unfold. The first dimension where the percentage drops to 0 (or very close to it) is chosen as the optimal embedding dimension for the data.

The calculation for False Nearest Neighbors can be represented by the following equation:

\eqn{\frac{||\hat{V}(t) - \hat{V}^{NN}t||^2 - || V(t) - V^{NN}(t)||^2} {||V(t) - V^{NN}(t)||} >R_{tol}}

Where \eqn{||V(t) – V(t)^{NN}||} is the distance between two vectors, and \eqn{||\hat{V}(t) – \hat{V}(t)^{NN}||} is the distance between the vector and its neighbor in d+1 dimensional space. If the vectors truly are neighbors, the results of \eqn{||\hat{V}(t) - \hat{V}^{NN}t||^2 - || V(t) - V^{NN}(t)||^2} should be constant. If a false neighbor is detected and is greater than \eqn{R_{tol} = 15} then the vector is determined to be a false neighbor.
}
\examples{
# Generate example time series data
x = rnorm(1000)

# Function parameters
maxDim = 10
delay = 1 # You can get this value like: ami_out$tau[1,1]
rtol = 10
atol = 15
fnn_tol = 0.01

# Compute false nearest neighbors
fnn_out = false_nearest_neighbors(x, maxDim = maxDim, delay = delay, rtol = rtol, 
                                  atol = atol, fnn_tol = fnn_tol)

}
\references{
Kennel, M. B., & Abarbanel, H. D. I. (2002). False neighbors and false strands: A reliable minimum embedding dimension algorithm. Phys. Rev. E, 66(2), 026209. https://doi.org/10.1103/PhysRevE.66.026209

Raffalt, P. C., Senderling, B., & Stergiou, N. (2020). Filtering affects the calculation of the largest Lyapunov exponent. Computers in Biology and Medicine, 122, 103786. https://doi.org/10.1016/j.compbiomed.2020.103786
}
