% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fgn_sim}
\alias{fgn_sim}
\title{Fractional Gaussian Noise Simulation}
\usage{
fgn_sim(n = 1000L, H = 0.7)
}
\arguments{
\item{n}{The length of the resulting time series}

\item{H}{The Hurst value of the resulting time series}
}
\description{
Function to generate Fractional Gaussian Noise
}
\details{
Create a time series with a specified Hurst value.
}
\examples{

ts_out = fgn_sim(n = 1000, H = 0.7)
}
