% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NONANr-package.R
\docType{package}
\name{nonanR-package}
\alias{nonanR}
\alias{nonanR-package}
\title{nonanR: Nonlinear Analysis Methods for Time Series Analysis}
\description{
This is a collection of functions that can be used for the nonlinear analysis of time series data. Analysis can be done by explicitly writing out code or using the Shiny app for a point-and-click analysis.
}
\author{
\strong{Maintainer}: Aaron D. Likens \email{alikens@unomaha.edu} (\href{https://orcid.org/0000-0002-6535-5772}{ORCID})

Authors:
\itemize{
  \item Joel H. Sommerfeld \email{joel.sommerfeld36@gmail.com} (\href{https://orcid.org/0000-0002-4972-0872}{ORCID})
  \item Seung Kyeom Kim \email{seungkyeomkim@unomaha.edu} (\href{https://orcid.org/0000-0001-5059-6867}{ORCID})
  \item Travis Wiltshire \email{T.J.Wiltshire@tilburguniversity.edu} (\href{https://orcid.org/0000-0001-7630-2695}{ORCID})
}

Other contributors:
\itemize{
  \item Tyler M. Wiles \email{tylerwiles@unomaha.edu} (\href{https://orcid.org/0000-0001-9915-6950}{ORCID}) [contributor]
}

}
\keyword{internal}
