% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fnn.R
\name{plot_fnn}
\alias{plot_fnn}
\title{False Nearest Neighbors Plot}
\usage{
plot_fnn(x)
}
\arguments{
\item{x}{The list object returned \code{fnn} in this package.}
}
\description{
Plot method for false nearest neighbors.
}
\examples{
# When performing phase space reconstruction there are two crucial steps. 
# - The first step is running the ami function. 

# Generate example time series data
x = fgn_sim(n = 1000, H = 0.9)

# Specify the parameters for ami
y = x
L = 50
bins = 30

# Run ami
ami_out = ami(x, y, L, bins)

# - The second step is running the fnn function
# Function parameters
maxDim = 10
delay = 3 # You can get this value like: ami_out$tau[1,1]
rtol = 10
atol = 15
fnn_tol = 0.01

# Run False Nearest Neighbors
fnn_out = false_nearest_neighbors(x, maxDim = maxDim, delay = delay, rtol = rtol, 
                                 atol = atol, fnn_tol = fnn_tol)

# Plot fnn_out
plot_fnn(fnn_out)

## Second example
# Generate example time series data
x = rnorm(1000)

# Function paramters
maxDim = 10
delay = 1 # You can get this value like: ami_out$tau[1,1]
rtol = 10
atol = 15
fnn_tol = 0.01

# Compute false nearest neighbors
fnn_out = false_nearest_neighbors(x, maxDim = maxDim, delay = delay, rtol = rtol, 
                                  atol = atol, fnn_tol = fnn_tol)

# Plot fnn_out
plot_fnn(fnn_out)

}
