% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fnn.R
\name{plot_fnn}
\alias{plot_fnn}
\title{False Nearest Neighbors Plot}
\usage{
plot_fnn(x)
}
\arguments{
\item{x}{The list object returned \code{fnn} in this package.}
}
\description{
Plot method for false nearest neighbors.
}
\examples{

# Create a time series and perform FNN algorithm
x = fgn_sim(n = 1000, H = 0.9)

tau = ami_out$tau[1,1]
mmax = 12
rtol = 15
atol = 2
L = 1

fnn_out = fnn(x = x, tau = tau, mmax = mmax, rtol = rtol, atol = atol)

# Plot fnn_out
plot_fnn(fnn_out)

}
