% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gui.R
\name{runGUI}
\alias{runGUI}
\title{This function allows you to run a shiny app that will do all the things for you}
\usage{
runGUI()
}
\description{
This function allows you to run a shiny app that will do all the things for you
}
