% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SampleEntropy}
\alias{SampleEntropy}
\title{Calute the sample entropy of a time series with this function.}
\usage{
SampleEntropy(x, m, R)
}
\arguments{
\item{x}{- a single column time series}

\item{m}{- the length of the vectors to be compared for matches}

\item{R}{- the radius for accepting matches}
}
\description{
Calute the sample entropy of a time series with this function.
}
\details{
Sample entropy can be used to quantify the complexity of a time series. A higher sample entropy value is associated with a more complex signal (more random). On the other hand a time series with a lower sample entropy value could be considered more predictable and repetitive.
}
\examples{

ts = rnorm(1000)

m = 2
R = 0.2

SE = SampleEntropy(ts, m, R)

}
\references{
Richman, J.S., Moorman, J.R., 2000. Physiological time-series analysis using approximate entropy and sample entropy. Am. J. Physiol. Heart Circ. Physiol. 278. https://doi.org/10.1152/ajpheart.2000.278.6.H2039
}
