% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SymbolicEntropy}
\alias{SymbolicEntropy}
\title{This function computes the symbolic entropy on a time series really, really quickly.}
\usage{
SymbolicEntropy(data, thresholdVal, seqLength)
}
\arguments{
\item{data}{- A vector of integers}

\item{thresholdVal}{- the threshold of the search that you want to do}

\item{seqLength}{- the length of the sequence that you want to find}
}
\description{
This function computes the symbolic entropy on a time series really, really quickly.
}
\details{
Like all entropy functions, this one also quantifies the amount of complexity (or uncertainty/unpredictability) in the signal. Symbolic entropy can be used to gather greater insight into the underlying patterns seen in movement data.
}
\examples{

ts = rnorm(1000)

thresholdVal = 2
seqLength = 0.2

SymE = SymbolicEntropy(ts, thresholdVal, seqLength)

}
\references{
Aziz, W., Arif, M. Complexity analysis of stride interval time series by threshold dependent symbolic entropy. Eur J Appl Physiol 98, 30–40 (2006). https://doi.org/10.1007/s00421-006-0226-5
}
