% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nonan.R
\name{theme_nonan}
\alias{theme_nonan}
\title{Theme to create NONAN style plots}
\usage{
theme_nonan(
  background = "#ffffff",
  main_text = "#000000",
  axis_text = "#444444",
  axes = "#D9D9D9"
)
}
\arguments{
\item{background}{Color for the background of the plot}

\item{main_text}{Color of the text for the main title and axis titles}

\item{axis_text}{Color of the axis tick labels}

\item{axes}{Color for the axis lines}
}
\description{
This function allows you to use the NONAN theme for ggplot2 graphics
}
\examples{

library(ggplot2)
 
iris_dat = iris
 
# Plot the data and add nonan_theme()
ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +
geom_point() +
labs(title = "Sepal Length vs Sepal Width",
     x = "Sepal Length",
     y = "Sepal Width") +
 theme_nonan()
  
  
# Alternatively you can provide explicit color values for the different features of the graph
ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +
geom_point() +
labs(title = "Sepal Length vs Sepal Width",
     x = "Sepal Length",
     y = "Sepal Width") +
 theme_nonan(background = "grey15", 
             main_text = "goldenrod", 
             axis_text = "maroon", 
             axes = "#ffffff")

}
