% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nonan.R
\name{theme_nonan}
\alias{theme_nonan}
\title{Theme to create NONAN style plots}
\usage{
theme_nonan(
  background = "#ffffff",
  main_text = "#000000",
  axis_text = "#444444",
  axes = "#D9D9D9"
)
}
\arguments{
\item{background}{Color for the background of the plot}

\item{main_text}{Color of the text for the main title and axis titles}

\item{axis_text}{Color of the axis tick labels}

\item{axes}{Color for the axis lines}
}
\description{
This function allows you to use the NONAN theme for ggplot2 graphics
}
\examples{

library(ggplot2)
library(NONANr)

# Extract data
your_column <- healthy_young$contact_lt
# Find the differences between consecutive elements
differences <- diff(your_column)
 
# Find the indices where the difference is 1000 and the previous value was 0
indices <- which(differences == 1000)
 
stride_time = diff(indices)/ 200 # 200 is the sampling frequency of the data
 
# Combine into a data frame for ggplot
dat = data.frame("stride_num" = 1:length(stride_time), "stride_time" = stride_time)
 
# Plot the data and add nonan_theme()
ggplot(dat, ggplot2::aes(x = stride_num, y = stride_time)) + 
 geom_line(color = "black", linewidth = 1) + 
 theme_nonan()
  
  
# Alternatively you could provide color values for the different features of the graph
ggplot(dat, ggplot2::aes(x = stride_num, y = stride_time)) + 
 geom_line(color = "black", linewidth = 1) + 
 theme_nonan(background = "#BCBBBA", main_text = "#D71920", axis_text = "#681820", axes = "#474648")

}
