% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nonan.R
\name{theme_nonan}
\alias{theme_nonan}
\title{This function allows you to use the NONAN theme for ggplot2 graphics}
\usage{
theme_nonan(background = "#646666", text = "#D9D9D9", axes = "#8A8A8A")
}
\description{
This function allows you to use the NONAN theme for ggplot2 graphics
}
\examples{

library(ggplot2)
library(NONANr)

# Extract data
your_column <- NONANr::healthy_young$`Contact RT`
# Find the differences between consecutive elements
differences <- diff(your_column)
 
# Find the indices where the difference is 1000 and the previous value was 0
indices <- which(differences == 1000)
 
stride_time = diff(indices)/ 200 # 200 is the sampling frequency of the data
 
# Combine into a data frame for ggplot
dat = data.frame("stride_num" = 1:length(stride_time), "stride_time" = stride_time)
 
# Plot the data and add nonan_theme()
ggplot(dat, ggplot2::aes(x = stride_num, y = stride_time)) + 
 geom_line(color = "black", linewidth = 1) + 
 theme_nonan()
  
  
# Alternatively you could provide color values for the different features of the graph
ggplot(dat, ggplot2::aes(x = stride_num, y = stride_time)) + 
 geom_line(color = "black", linewidth = 1) + 
 theme_nonan(background = "#ffffff", text = "#000000", axes = "#D9D9D9")

}
